# perf: use faster fasta parser implementation