--- a/packages/nextalign/CMakeLists.txt
+++ b/packages/nextalign/CMakeLists.txt
@@ -33,6 +33,7 @@ add_library(${PROJECT_NAME} STATIC
   src/io/gene.io.h
   src/io/parseGeneMapGff.cpp
   src/io/parseSequences.cpp
+  src/io/parseSequencesSlow.cpp
   src/match/matchAa.cpp
   src/match/matchAa.h
   src/match/matchNuc.cpp