--- a/packages/nextclade_wasm/src/main.cpp
+++ b/packages/nextclade_wasm/src/main.cpp
@@ -190,7 +190,7 @@ public:
 
 AlgorithmInput parseRefSequence(const std::string& refFastaStr, const std::string& refFastaName) {
   std::stringstream refFastaStream{refFastaStr};
-  const auto parsed = parseSequences(refFastaStream, refFastaName);
+  const auto parsed = parseSequencesSlow(refFastaStream, refFastaName);
 
   if (parsed.size() != 1) {
     throw std::runtime_error(fmt::format("Error: {:d} sequences found in reference sequence file ('{:s}'), expected 1",
@@ -220,9 +220,10 @@ std::string parsePcrPrimerCsvRowsStr(const std::string& pcrPrimersStr, const std
 void parseSequencesStreaming(const std::string& queryFastaStr, const std::string& queryFastaName,
   const emscripten::val& onSequence, const emscripten::val& onComplete) {
   std::stringstream queryFastaStringstream{queryFastaStr};
-  auto inputFastaStream = makeFastaStream(queryFastaStringstream, queryFastaName);
-  while (inputFastaStream->good()) {
-    onSequence(inputFastaStream->next());
+  auto inputFastaStream = makeFastaStreamSlow(queryFastaStringstream, queryFastaName);
+  AlgorithmInput input;
+  while (inputFastaStream->next(input)) {
+    onSequence(std::move(input));
   }
   onComplete();
 }