… packages/nextalign/tests/parseSequences.test.cpp