This improves runtime performance of Nextalign and Nextclade CLI.

The speedup is achieved by using a faster FASTA parser. As a base I used code from kseqpp from Ali Ghaffaari (`@cartoonist`): https://github.com/cartoonist/kseqpp


We are loosing these features:

 - Sequence names are no longer tracked and deduplicated. It required storing a map of all sequence names and lookups into this map on each new sequence. So a map of 5M items for GISAID. This feature is a  "nice-to-have" and I think we can sacrifice it.
 - Plain text files are not supported by this parser.
 - Hard to test, because it uses lower-level C APIs to read files and parsing logic intertwined with I/O. Old tests are commented out for now.
 - Requires UNIX-specifics (`unistd.h` and `fcntl.h`). But we only ship Linux and MacOS, so this should be fine. If Windows release will be needed, we can use a second implementation there conditionally.


### Benchmarks

Setup:

 - AWS EC2 c5.24xlarge (96 cores @ 3.6 GHz, 192GiB RAM)
 - SSD: gp2 1TB 3000 IOPS
 - Ubuntu 20.04
 - 100k randomly sampled GISAID sequences

Results:

```
|-------------|---------------------|---------------------|---------|
|             |   Wall clock time   |  Throughput, seq/s  |         |
|-------------|-----------|---------|-----------|---------|---------|
|             |   before  |   after |   before  |  after  | speedup |
|-------------|-----------|---------|-----------|---------|---------|
| Nextalign   |   5m 09s  |     51s |      324  |    1933 |   6.05x |
| Nextclade   |   6m 15s  |  3m 47s |      266  |     440 |   1.65x |
|-------------|-----------|---------|-----------|---------|---------|

```

### Scaling

Both tools now scale nicely to a large number of cores.

Before:
![before](https://user-images.githubusercontent.com/9403403/144193574-39a307ef-4c1f-440d-9575-5c313355213d.png)

After:
![after](https://user-images.githubusercontent.com/9403403/144193626-7560d6b5-a03d-4262-abec-cbc64e8ac3e2.png)


### Try it:

Staging CLI executables are under "Artifacts" tab on the page for each architecture:
https://app.circleci.com/pipelines/github/nextstrain/nextclade/1873/workflows/d8c194a3-6dbe-4f7b-929a-5a2433b1c470


### Repro

Prepare sample data:

```bash
seqkit sample --proportion=0.02 gisaid.fasta | seqkit head --number=100000 -o gisaid-100k.fasta
seqkit stats gisaid-100k.fasta
```

Alias for time measurement:

```bash
alias TIME="/usr/bin/time -f 'Wall clock time:         %E\nMaximum resident memory: %M KB\n'"
```

Get executables:

```bash

curl -fsSL https://github.com/nextstrain/nextclade/releases/download/1.5.1/nextalign-Linux-x86_64 -o nextalign-1.5.1

curl -fsSL https://github.com/nextstrain/nextclade/releases/download/1.5.1/nextclade-Linux-x86_64 -o nextclade-1.5.1

curl -fsSL https://1078-271799467-gh.circle-artifacts.com/0/nextalign-Linux-x86_64 -o nextalign-faster

curl -fsSL https://1078-271799467-gh.circle-artifacts.com/0/nextclade-Linux-x86_64 -o nextclade-faster

```


Get dataset:

```bash
nextclade dataset get --name=sars-cov-2 --output-dir=data/
```

Run:

```bash
reset; TIME ./nextalign-1.5.1 --sequences=gisaid-100k.fasta --reference=data/reference.fasta --genemap=data/genemap.gff --genes=E,M,N,ORF1a,ORF1b,ORF3a,ORF6,ORF7a,ORF7b,ORF8,ORF9b,S --output-dir=tmp/ --output-basename=nextalign --include-reference --silent

reset; TIME ./nextalign-faster --sequences=gisaid-100k.fasta --reference=data/reference.fasta --genemap=data/genemap.gff --genes=E,M,N,ORF1a,ORF1b,ORF3a,ORF6,ORF7a,ORF7b,ORF8,ORF9b,S --output-dir=tmp/ --output-basename=nextalign --include-reference --silent

reset; TIME ./nextclade-1.5.1 --input-fasta=gisaid-100k.fasta --input-dataset=data/ --output-tsv=tmp/nextclade.tsv --output-dir=tmp/ --output-basename=nextclade --include-reference --silent

reset; TIME ./nextclade-faster --input-fasta=gisaid-100k.fasta --input-dataset=data/ --output-tsv=tmp/nextclade.tsv --output-dir=tmp/ --output-basename=nextclade --include-reference --silent
```