@ivan-aksamentov 2021-12-01
I ran some additional numbers, to see how it scales with number of threads:


[bench.csv](https://github.com/nextstrain/nextclade/files/7633582/bench.csv)
[bench.sh.zip](https://github.com/nextstrain/nextclade/files/7633587/bench.sh.zip)


![01](https://user-images.githubusercontent.com/9403403/144231097-e77741f6-f07f-42d7-a43e-854138f01b69.png)
![02](https://user-images.githubusercontent.com/9403403/144231105-f77d282c-17c9-4862-bf47-054cc0c2e8ba.png)