@corneliusroemer 2021-12-01
Amazing! Finally the fasta parsing bottleneck is removed! Same graph for
Nextclade would be interesting too.

On Wed, Dec 1, 2021, 13:03 Ivan Aksamentov ***@***.***> wrote:

> I ran some additional numbers, to see how it scales with number of threads:
>
> bench.csv
> <https://github.com/nextstrain/nextclade/files/7633582/bench.csv>
> bench.sh.zip
> <https://github.com/nextstrain/nextclade/files/7633587/bench.sh.zip>
>
> [image: 01]
> <https://user-images.githubusercontent.com/9403403/144231097-e77741f6-f07f-42d7-a43e-854138f01b69.png>
> [image: 02]
> <https://user-images.githubusercontent.com/9403403/144231105-f77d282c-17c9-4862-bf47-054cc0c2e8ba.png>
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/nextstrain/nextclade/pull/632#issuecomment-983574365>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AF77AQN6PLPBUL6V6MQFB3DUOYFJRANCNFSM5JDZSZYA>
> .
>