@ivan-aksamentov 2021-12-02
Nextclade's improvement is less impressive. But I have a few ideas for later.
![01](https://user-images.githubusercontent.com/9403403/144412036-4b0e6028-ac61-4c85-93ad-948c8eeeb521.png)