@ivan-aksamentov 2021-12-02
I brought back the slower parser, because we cannot use the new parser (which can only read from file) in the WebAssembly module (which only has string inputs). Later we might figure out how to adapt the new parser to string inputs.