@ivan-aksamentov 2021-12-02
I just verified on 100k sequences that both pairs of tools produce byte-identical results. So this PR and executables are ready to be reviewed.