@rneher 2021-12-05
I just gave this a go and it worked fine. I reproduce the `nextalign` speed up with 8 threads.