@ivan-aksamentov 2021-12-08
I verified once again that this is stable on 1M sequence dataset, and also benchmarked against v1.6.0:

```
experiment_name,problem_size,num_jobs,wall_seconds,max_mem_kb
nextalign-1.6.0,100000,96,315.11,10548
nextalign-1.6.0,100000,80,315.09,10460
nextalign-1.6.0,100000,64,315.67,10516
nextalign-1.6.0,100000,48,315.52,10516
nextalign-1.6.0,100000,32,315.56,10492
nextalign-1.6.0,100000,16,316.78,10556
nextalign-1.6.0,100000,8,429.80,10492
nextalign-faster,100000,96,65.03,10628
nextalign-faster,100000,80,65.05,10572
nextalign-faster,100000,64,66.80,10548
nextalign-faster,100000,48,75.02,10560
nextalign-faster,100000,32,100.20,10544
nextalign-faster,100000,16,186.60,10632
nextalign-faster,100000,8,346.59,10460
nextclade-1.6.0,100000,96,389.74,10488
nextclade-1.6.0,100000,80,392.52,10540
nextclade-1.6.0,100000,64,387.37,10404
nextclade-1.6.0,100000,48,388.20,10380
nextclade-1.6.0,100000,32,478.83,10516
nextclade-1.6.0,100000,16,827.01,10460
nextclade-1.6.0,100000,8,1541.98,10456
nextclade-faster,100000,96,220.94,10444
nextclade-faster,100000,80,239.25,10556
nextclade-faster,100000,64,263.26,10572
nextclade-faster,100000,48,313.00,10468
nextclade-faster,100000,32,417.64,10552
nextclade-faster,100000,16,760.09,10444
nextclade-faster,100000,8,1435.72,10380
```

Merging